pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lead-search-toolkit"
version = "0.1.0"
description = "Config-driven lead search toolkit — find potential customers via Google dorks, LinkedIn, and Israeli gov data"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Sergei Benkovitch", email = "hello@nomemoo.com"}]

dependencies = [
    "ddgs>=6.0",
    "requests>=2.28",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
linkedin = ["linkedin-api>=2.0"]

[project.scripts]
lead-search = "lead_search.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
